Split DomainObject into smaller mixins
